batch renderer debuging

+++ Space Explorers/Texture.h
@@ -8,8 +8,8 @@ class Texture
 public:
 	Texture(std::string path);
 	Texture(int width, int height, glm::u8vec4 fillColor);
-	void set(int x, int y, glm::u8vec4 color);
-	glm::u8vec4 get(int x, int y);
+	void set(size_t x, size_t y, glm::u8vec4 color);
+	glm::u8vec4 get(size_t x, size_t y);
 	int getWidth();
 	int getHeight();
 	glm::ivec2 getSize();
